fix: close portal verifier review gaps
