+++ src/travel_plan_permission/http_service.py
@@ -24,11 +24,17 @@ from fastapi import (
 )
 from fastapi.responses import HTMLResponse, RedirectResponse
 from fastapi.templating import Jinja2Templates
-from pydantic import BaseModel, Field, ValidationError
+from pydantic import BaseModel, Field
 
-from .canonical import CanonicalTripPlan, canonical_trip_plan_to_model
+from .canonical import CanonicalTripPlan
 from .models import TripPlan
 from .planner_auth import PlannerAuthConfig, authenticate_request
+from .portal_review import (
+    PortalArtifact,
+    PortalReviewState,
+    portal_review_state,
+    portal_validation_state,
+)
 from .policy_api import (
     PlannerPolicySnapshot,
     PlannerPolicySnapshotRequest,
@@ -41,10 +47,8 @@ from .policy_api import (
     get_evaluation_result,
     get_policy_snapshot,
     poll_execution_status,
-    render_travel_spreadsheet_bytes,
     submit_proposal,
 )
-from .prompt_flow import build_output_bundle, generate_questions, required_field_gaps
 from .review_workflow import ReviewAction, ReviewRequest, ReviewWorkflowStore
 from .security import Permission
 
@@ -156,33 +160,6 @@ class PortalDraft:
     answers: dict[str, object]
     updated_at: datetime
     cached_artifacts: dict[str, PortalArtifact] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class PortalArtifact:
-    """Generated portal artifact metadata and payload."""
-
-    filename: str
-    content: bytes
-    media_type: str
-
-
-@dataclass(frozen=True)
-class PortalReviewState:
-    """Computed portal review context derived from a draft."""
-
-    draft_id: str
-    answers: dict[str, object]
-    missing_fields: list[str]
-    next_questions: list[dict[str, object]]
-    validation_errors: list[str]
-    canonical_payload: dict[str, object] | None
-    trip_plan: TripPlan | None
-    policy_snapshot: PlannerPolicySnapshot | None
-    policy_result: Any | None
-    artifacts: dict[str, PortalArtifact]
-    submission_response: PlannerProposalOperationResponse | None = None
-    manager_review: ReviewRequest | None = None
 
 
 class PlannerRuntimeConfigError(RuntimeError):
@@ -560,125 +537,6 @@ def _canonical_payload_from_answers(answers: dict[str, object]) -> dict[str, obj
     return payload
 
 
-def _review_validation_errors(exc: ValidationError) -> list[str]:
-    errors: list[str] = []
-    for error in exc.errors():
-        location = ".".join(str(part) for part in error["loc"])
-        errors.append(f"{location}: {error['msg']}")
-    return errors
-
-
-def _portal_artifacts(
-    *,
-    canonical: CanonicalTripPlan,
-    plan: TripPlan,
-    answers: dict[str, object],
-) -> dict[str, PortalArtifact]:
-    itinerary_excel = render_travel_spreadsheet_bytes(plan, canonical_plan=canonical)
-    bundle = build_output_bundle(itinerary_excel=itinerary_excel, answers=answers)
-    itinerary_payload = bundle["itinerary_excel"]
-    summary_payload = bundle["summary_pdf"]
-    if not isinstance(itinerary_payload, dict):
-        raise RuntimeError("itinerary_excel bundle payload must be a mapping")
-    if not isinstance(summary_payload, dict):
-        raise RuntimeError("summary_pdf bundle payload must be a mapping")
-    return {
-        "itinerary": PortalArtifact(
-            filename=str(itinerary_payload["filename"]),
-            content=itinerary_excel,
-            media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-        ),
-        "summary": PortalArtifact(
-            filename=str(summary_payload["filename"]),
-            content=bytes(summary_payload["content"]),
-            media_type=str(summary_payload["mime_type"]),
-        ),
-    }
-
-
-def _portal_validation_state(answers: dict[str, object]) -> PortalReviewState:
-    missing_fields = required_field_gaps(
-        answers, required_fields=_PORTAL_REQUIRED_FIELDS
-    )
-    next_questions: list[dict[str, object]] = [
-        {
-            "prompt": question.prompt,
-            "fields": ", ".join(question.fields),
-            "kind": question.kind,
-        }
-        for question in generate_questions(answers, max_questions=4)
-    ]
-    validation_errors: list[str] = []
-    canonical_payload: dict[str, object] | None = None
-
-    if not missing_fields:
-        canonical_payload = _canonical_payload_from_answers(answers)
-        try:
-            CanonicalTripPlan.model_validate(canonical_payload)
-        except ValidationError as exc:
-            validation_errors = _review_validation_errors(exc)
-
-    return PortalReviewState(
-        draft_id="",
-        answers=answers,
-        missing_fields=missing_fields,
-        next_questions=next_questions,
-        validation_errors=validation_errors,
-        canonical_payload=canonical_payload,
-        trip_plan=None,
-        policy_snapshot=None,
-        policy_result=None,
-        artifacts={},
-    )
-
-
-def _portal_review_state(
-    draft_id: str,
-    answers: dict[str, object],
-    *,
-    submission_response: PlannerProposalOperationResponse | None = None,
-    manager_review: ReviewRequest | None = None,
-) -> PortalReviewState:
-    validation_state = _portal_validation_state(answers)
-    missing_fields = validation_state.missing_fields
-    next_questions = validation_state.next_questions
-    validation_errors = validation_state.validation_errors
-    canonical_payload = validation_state.canonical_payload
-    trip_plan: TripPlan | None = None
-    policy_snapshot: PlannerPolicySnapshot | None = None
-    policy_result: Any | None = None
-    artifacts: dict[str, PortalArtifact] = {}
-
-    if not missing_fields and canonical_payload is not None and not validation_errors:
-        canonical = CanonicalTripPlan.model_validate(canonical_payload)
-        trip_plan = canonical_trip_plan_to_model(canonical)
-        policy_snapshot = get_policy_snapshot(
-            trip_plan,
-            PlannerPolicySnapshotRequest(trip_id=trip_plan.trip_id),
-        )
-        policy_result = check_trip_plan(trip_plan)
-        artifacts = _portal_artifacts(
-            canonical=canonical,
-            plan=trip_plan,
-            answers=answers,
-        )
-
-    return PortalReviewState(
-        draft_id=draft_id,
-        answers=answers,
-        missing_fields=missing_fields,
-        next_questions=next_questions,
-        validation_errors=validation_errors,
-        canonical_payload=canonical_payload,
-        trip_plan=trip_plan,
-        policy_snapshot=policy_snapshot,
-        policy_result=policy_result,
-        artifacts=artifacts,
-        submission_response=submission_response,
-        manager_review=manager_review,
-    )
-
-
 def _portal_template_context(
     request: Request,
     review: PortalReviewState | None = None,
@@ -752,7 +610,11 @@ def create_app(store: PlannerProposalStore | None = None) -> FastAPI:
     @app.post("/portal/draft")
     async def portal_draft_review(request: Request) -> Response:
         answers = _portal_answers_from_encoded_body(await request.body())
-        review = _portal_validation_state(answers)
+        review = portal_validation_state(
+            answers,
+            required_fields=_PORTAL_REQUIRED_FIELDS,
+            canonical_payload_builder=_canonical_payload_from_answers,
+        )
         if review.missing_fields or review.validation_errors:
             return _TEMPLATES.TemplateResponse(
                 request=request,
@@ -780,9 +642,11 @@ def create_app(store: PlannerProposalStore | None = None) -> FastAPI:
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail=f"No portal draft found for '{draft_id}'.",
             )
-        review = _portal_review_state(
+        review = portal_review_state(
             draft.draft_id,
             draft.answers,
+            required_fields=_PORTAL_REQUIRED_FIELDS,
+            canonical_payload_builder=_canonical_payload_from_answers,
             manager_review=proposal_store.lookup_manager_review_for_draft(
                 draft.draft_id
             ),
@@ -811,7 +675,12 @@ def create_app(store: PlannerProposalStore | None = None) -> FastAPI:
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail=f"No portal draft found for '{draft_id}'.",
             )
-        review = _portal_review_state(draft.draft_id, draft.answers)
+        review = portal_review_state(
+            draft.draft_id,
+            draft.answers,
+            required_fields=_PORTAL_REQUIRED_FIELDS,
+            canonical_payload_builder=_canonical_payload_from_answers,
+        )
         if (
             review.trip_plan is None
             or review.missing_fields
@@ -838,9 +707,11 @@ def create_app(store: PlannerProposalStore | None = None) -> FastAPI:
             submission_response,
         )
         manager_review = proposal_store.create_manager_review(review)
-        review = _portal_review_state(
+        review = portal_review_state(
             draft.draft_id,
             draft.answers,
+            required_fields=_PORTAL_REQUIRED_FIELDS,
+            canonical_payload_builder=_canonical_payload_from_answers,
             submission_response=submission_response,
             manager_review=manager_review,
         )
@@ -966,7 +837,12 @@ def create_app(store: PlannerProposalStore | None = None) -> FastAPI:
             )
         artifacts = draft.cached_artifacts
         if not artifacts:
-            review = _portal_review_state(draft.draft_id, draft.answers)
+            review = portal_review_state(
+                draft.draft_id,
+                draft.answers,
+                required_fields=_PORTAL_REQUIRED_FIELDS,
+                canonical_payload_builder=_canonical_payload_from_answers,
+            )
             artifacts = review.artifacts
             if artifacts:
                 proposal_store.cache_portal_artifacts(draft.draft_id, artifacts)
